move `AssertProcessEvents` to a test only file
